fix(static-analysis): resolve all PHPStan errors across packages and tests
- Fix Filament namespace changes (Forms → Schemas) in 12 Modern* schemas and AI actions
- Fix MigrationFilesystemInterface → MigrationFileManagerInterface in address/assistant/mosaic install commands
- Add phpstan/stubs for forward-reference interfaces (PageHeaderActionExtender, SiteHeaderActionExtender, BelongsToWorkspace) using scanFiles
- Exclude dead OpenAIProviderTest (references removed class) from analysis
- Fix method.childParameterType: use mixed $app in TestCase overrides to match Testbench parent signature
- Remove redundant casts (int/float/string/array) on typed properties and config() calls
- Fix SectionRegistry: replace empty(), short ternary ?:, and mixed ternary condition
- Add @property docblock to AiCreatorSession for dynamically-set status field
- Type all untyped closure/callback parameters across plugins and test files
- Ignore forward-reference classes (WorkspaceRegistry, PageTranslation) and dynamic relation (layoutWidgets)

+++ packages/assistant/src/Support/Pipelines/SuggestMetaDescriptionsPipeline.php
@@ -77,7 +77,7 @@ class SuggestMetaDescriptionsPipeline
                 ['role' => 'system', 'content' => (string) ($prompt['system'] ?? '')],
                 ['role' => 'user', 'content' => $userMessage . "\nPlease provide 3 meta description options as a simple bullet list."],
             ],
-            'max_tokens' => (int) config('capell-assistant.prism.max_tokens', 128),
+            'max_tokens' => config('capell-assistant.prism.max_tokens', 128),
             'temperature' => 0.7,
         ];
 
@@ -112,8 +112,8 @@ class SuggestMetaDescriptionsPipeline
                 'output' => implode("\n", (array) ($payload['result'] ?? [])),
                 'prompt_tokens' => (int) ($response->metadata['prompt_tokens'] ?? 0),
                 'completion_tokens' => (int) ($response->metadata['completion_tokens'] ?? 0),
-                'total_tokens' => (int) $response->tokensUsed,
-                'duration' => (float) $response->duration,
+                'total_tokens' => $response->tokensUsed,
+                'duration' => $response->duration,
                 'metadata' => $response->metadata,
             ]);
         }
